Bug fixes, improvements and documentation updates

+++ modules/audit_nis_entries.sh
@@ -21,33 +21,36 @@ audit_nis_entries () {
     for check_file in /etc/passwd /etc/shadow /etc/group; do
       if test -r "$check_file"; then
         if [ "$audit_mode" != 2 ]; then
-          file_entries=$( grep "^\+" "$check_file" )
-          for file_entry in $file_entries; do
-            if [ "$audit_mode" = 1 ]; then
-              increment_insecure "NIS entry \"$file_entry\" in $check_file"
-              verbose_message    "sed -e \"s/^\+/#&/\" < $check_file > $temp_file" "fix"
-              verbose_message    "cat $temp_file > $check_file" "fix"
-            fi
-            if [ "$audit_mode" = 0 ]; then
-              backup_file $check_file
-              verbose_message "File \"$check_file\" to have no NIS entries" "set"
-              sed -e "s/^+/#&/" < "$check_file" > "$temp_file"
-              cat "$temp_file" > "$check_file"
-              if [ "$os_name" = "SunOS" ]; then
-                if [ "$os_version" != "11" ]; then
-                  pkgchk -f -n -p "$check_file" 2> /dev/null
-                else
-                  pkg fix $( pkg search "$check_file" | grep pkg | awk '{print $4}' )
+          entry_check=$( grep "^\+" "$check_file" | wc -l)
+          if [ ! "$entry_check" = "0" ]; then
+            file_entries=$( grep "^\+" "$check_file" )
+            for file_entry in $file_entries; do
+              if [ "$audit_mode" = 1 ]; then
+                increment_insecure "NIS entry \"$file_entry\" in $check_file"
+                verbose_message    "sed -e \"s/^\+/#&/\" < $check_file > $temp_file" "fix"
+                verbose_message    "cat $temp_file > $check_file" "fix"
+              fi
+              if [ "$audit_mode" = 0 ]; then
+                backup_file $check_file
+                verbose_message "File \"$check_file\" to have no NIS entries" "set"
+                sed -e "s/^+/#&/" < "$check_file" > "$temp_file"
+                cat "$temp_file" > "$check_file"
+                if [ "$os_name" = "SunOS" ]; then
+                  if [ "$os_version" != "11" ]; then
+                    pkgchk -f -n -p "$check_file" 2> /dev/null
+                  else
+                    pkg fix $( pkg search "$check_file" | grep pkg | awk '{print $4}' )
+                  fi
+                fi
+                if [ -f "$temp_file" ]; then
+                  rm "$temp_file"
                 fi
               fi
-              if [ -f "$temp_file" ]; then
-                rm "$temp_file"
+            done
+            if [ "$file_entry" = "" ]; then
+              if [ "$audit_mode" = 1 ]; then
+                increment_secure "No NIS entries in \"$check_file\""
               fi
-            fi
-          done
-          if [ "$file_entry" = "" ]; then
-            if [ "$audit_mode" = 1 ]; then
-              increment_secure "No NIS entries in \"$check_file\""
             fi
           fi
         else
